new config: moved classes out of __init__.py
